Enable context receivers
